requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.8
djangify==1.0.0
Django==3.2.9
django-ckeditor==6.1.0
django-galleryfield==1.3.dev0
django-js-asset==1.2.2
django-npm==1.0.0
django-paypal==1.1.1
idna==3.3
npm==0.1.1
optional-django==0.1.0
Pillow==7.1.1
pycryptodome==3.11.0
pytz==2021.3
requests==2.26.0
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
typing_extensions==4.0.0
urllib3==1.26.7
paypal-checkout-serversdk
razorpay